+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - LoRaCloud DAS integration page in the Console.
+- User Agent metadata on published events (when available).
 
 ### Changed